// src/app-types.ts
interface Position {
  id: number;
  title: string;
}
// ...
export interface Visit {
  id: number;
  clients_count: number;
  date: string;
  datetime: string;
  create_date: string;
  comment: string;
  deleted: boolean;
  attendance: number;
  length: number;
  notify_by_sms: number;
  notify_by_email: number;
  master_requested: boolean;
  online: boolean;
  api_id: string;
  last_change_date: string;
  prepaid: boolean;
  prepaid_confirmed: boolean;
  activity_id: number;
  paid_amount: number;
  allow_delete_record: boolean;
  allow_change_record: boolean;
  is_confirmation_needed: boolean;
  review_enabled: boolean;
  token: string;
  company: VisitCompany;
  resource_instances: any[];
  charge: any[];
  staff: Staff;
  services: VisitService[];
}
interface VisitService {
  id: number;
  title: string;
  cost: number;
  price_min: number;
  price_max: number;
  discount: number;
  amount: number;
  seance_length: number;
  api_id: string;
  abonement_restriction: number;
}

interface Staff {
  id: number;
  name: string;
  specialization: string;
  show_rating: number;
  rating: number;
  votes_count: number;
  avatar: string;
  avatar_big: string;
  comments_count: number;
  is_deleted: boolean;
  position: Position;
}
interface Position {
  id: number;
  chain_id: number;
  title: string;
  description: string;
  services_binding_type: number;
  rules_required_fields: any[];
  only_chain_appointment: boolean;
}
interface VisitCompany {
  id: number;
  title: string;
  public_title: string;
  country_id: number;
  country: string;
  city_id: number;
  city: string;
  phone: string;
  phones: string[];
  timezone: number;
  address: string;
  coordinate_lat: number;
  coordinate_lon: number;
  allow_delete_record: boolean;
  allow_change_record: boolean;
  allow_change_prepaid_record: boolean;
  allow_delete_prepaid_record: boolean;
  site: string;
  currency_short_title: string;
  allow_change_record_delay_step: number;
  allow_delete_record_delay_step: number;
  logo: string;
}